Move TestEdge to org.jgrapht.graph package.
- That's where the DefaultEdge class is defined.
- See https://github.com/jgrapht/jgrapht/pull/197#discussion_r61621398.
